Move addition into static method and add unit test
